just re-use the existing page fetch
