+++ changedetectionio/content_fetcher.py
@@ -7,6 +7,7 @@ from selenium.webdriver.common.proxy import Proxy as SeleniumProxy
 from selenium.common.exceptions import WebDriverException
 import requests
 import time
+import json
 import urllib3.exceptions
 
 
@@ -26,6 +27,102 @@ class Fetcher():
     headers = None
 
     fetcher_description ="No description"
+    xpath_element_js="""               
+                // Include the getXpath script directly, easier than fetching
+                !function(e,n){"object"==typeof exports&&"undefined"!=typeof module?module.exports=n():"function"==typeof define&&define.amd?define(n):(e=e||self).getXPath=n()}(this,function(){return function(e){var n=e;if(n&&n.id)return'//*[@id="'+n.id+'"]';for(var o=[];n&&Node.ELEMENT_NODE===n.nodeType;){for(var i=0,r=!1,d=n.previousSibling;d;)d.nodeType!==Node.DOCUMENT_TYPE_NODE&&d.nodeName===n.nodeName&&i++,d=d.previousSibling;for(d=n.nextSibling;d;){if(d.nodeName===n.nodeName){r=!0;break}d=d.nextSibling}o.push((n.prefix?n.prefix+":":"")+n.localName+(i||r?"["+(i+1)+"]":"")),n=n.parentNode}return o.length?"/"+o.reverse().join("/"):""}});
+                //# sourceMappingURL=index.umd.js.map             
+
+            
+                const findUpTag = (el) => {
+                  let r = el
+                  chained_css = [];
+            
+                  while (r.parentNode) {
+            
+                    if(r.classList.length >0) {
+                     // limit to just using 2 class names of each, stops from getting really huge selector strings
+                      current_css='.'+Array.from(r.classList).slice(0, 2).join('.');
+                      chained_css.unshift(current_css);
+            
+                      var f=chained_css.join(' ');
+                      var q=document.querySelectorAll(f);
+                      if(q.length==1) return current_css;
+                      if(f.length >120) return null;
+                    }  
+                    r = r.parentNode;
+                  }
+                  return null;
+                }
+
+                
+                var elements = document.getElementsByTagName("*");
+                var size_pos=[];
+                // after page fetch, inject this JS
+                // build a map of all elements and their positions (maybe that only include text?)
+                var bbox;
+                for (var i = 0; i < elements.length; i++) {   
+                 bbox = elements[i].getBoundingClientRect();
+                
+                 // forget reallysmall ones
+                 if (bbox['width'] <10 && bbox['height'] <10 ) {
+                   continue;
+                 }
+                
+                 // @todo the getXpath kind of sucks, it doesnt know when there is for example just one ID sometimes
+                 // it should not traverse when we know we can anchor off just an ID one level up etc..
+                 // maybe, get current class or id, keep traversing up looking for only class or id until there is just one match 
+                
+                 // 1st primitive - if it has class, try joining it all and select, if theres only one.. well thats us.
+                 xpath_result=false;
+                 try {
+                   var d= findUpTag(elements[i]);
+                   if (d) {
+                     xpath_result =d;
+                   }                
+                 } catch (e) {
+                   var x=1;
+                 }
+                
+                 // default back to the less intelligent one
+                 if (!xpath_result) {
+                   xpath_result = getXPath(elements[i]);                   
+                 } 
+                
+                 size_pos.push({
+                   xpath: xpath_result,
+                   width: bbox['width'], 
+                   height: bbox['height'],
+                   left: bbox['left'],
+                   top: bbox['top'],
+                   childCount: elements[i].childElementCount
+                 });                 
+                }
+                
+                
+                // inject the current one set in the css_filter, which may be a CSS rule
+                // used for displaying the current one in VisualSelector, where its not one we generated.
+                if (css_filter.length) {
+                   // is it xpath?
+                   if (css_filter.startsWith('/') ) {
+                     q=document.evaluate(css_filter, document, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null).singleNodeValue;
+                   } else {
+                     q=document.querySelector(css_filter);
+                   }
+                   if (q) {
+                       bbox = q.getBoundingClientRect();
+                       size_pos.push({
+                           xpath: css_filter,
+                           width: bbox['width'], 
+                           height: bbox['height'],
+                           left: bbox['left'],
+                           top: bbox['top'],
+                           childCount: q.childElementCount
+                         });
+                     }
+                }
+                
+                return size_pos;
+    """
 
     @abstractmethod
     def get_error(self):
@@ -58,6 +155,11 @@ class Fetcher():
     # Return true/false if this checker is ready to run, in the case it needs todo some special config check etc
     def is_ready(self):
         return True
+
+    @abstractmethod
+    def get_xpath_data(self, current_css_xpath_filter):
+        return None
+
 
 #   Maybe for the future, each fetcher provides its own diff output, could be used for text, image
 #   the current one would return javascript output (as we use JS to generate the diff)
@@ -162,6 +264,15 @@ class html_webdriver(Fetcher):
         # driver.quit() seems to cause better exceptions
         self.quit()
         return True
+
+    def get_xpath_data(self, current_css_xpath_filter):
+
+        # lazy quoting, probably going to be bad later.
+        css_filter = current_css_xpath_filter.replace('"', '\\"')
+        css_filter = css_filter.replace('\'', '\\\'')
+        info = self.driver.execute_script("var css_filter='{}';".format(css_filter)+self.xpath_element_js)
+        return info
+
 
     def quit(self):
         if self.driver:
